requirements.txt
pandas
numpy
streamlit
plotly
metalog
matplotlib
scipy
pyyaml